Relocate spaces API interfaces to to their own file and colocate shims for better organisation.
